rqt_rviz: Persisting rviz configuration in plugin (instance) settings.
This changes the behavior of rqt_rviz to load the state of rviz, as
it was when the perspective was closed, rather then starting with the
default configuration, which seems to be more intuitive, as other plugins,
e.g. the Image View behave the same way.

This allows users of this plugin to export a rqt perspective
which includes the current configuration of rviz.
Before this change there was no possibility to include a
rviz configuration in the rqt perspective file which meant that the
configuration needed to be loaded from the rviz menu after starting
rqt with a certain perspective.

It is saved in the instance settings, to allow users to have
multiple rviz configurations in one perspective.